Separate test program and convert to library
