+++ SrslBytecodeVmAndCodeGenerator.sln
@@ -1,6 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32210.238
+MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SrslBytecodeVmAndCodeGenerator", "SrslBytecodeVmAndCodeGenerator\SrslBytecodeVmAndCodeGenerator.csproj", "{2B941484-FB5E-4B3F-AF7D-BF61238D6F16}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestApp", "TestApp\TestApp.csproj", "{9577FEF2-3D0B-4CC3-9FF6-0FD56F564534}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -12,5 +17,15 @@ Global
 		{2B941484-FB5E-4B3F-AF7D-BF61238D6F16}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2B941484-FB5E-4B3F-AF7D-BF61238D6F16}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2B941484-FB5E-4B3F-AF7D-BF61238D6F16}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9577FEF2-3D0B-4CC3-9FF6-0FD56F564534}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9577FEF2-3D0B-4CC3-9FF6-0FD56F564534}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9577FEF2-3D0B-4CC3-9FF6-0FD56F564534}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9577FEF2-3D0B-4CC3-9FF6-0FD56F564534}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {38A55B41-EE01-4DA7-97AF-ECBE617D3D8E}
 	EndGlobalSection
 EndGlobal
